change from float do Decimal

+++ transaction.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 
 @dataclass
 class Transaction:
@@ -7,20 +8,20 @@ class Transaction:
     pair: str
     base_currency: str
     quote_currency: str
-    price: float
+    price: Decimal  # Changed from float to Decimal
     time: datetime
     ordertxid: str
     aclass: str
     maker: bool
     trade_id: str
-    vol: float
+    vol: Decimal    # Should also be Decimal
     ordertype: str
-    cost: float
-    fee: float
+    cost: Decimal   # Should also be Decimal
+    fee: Decimal    # Should also be Decimal
     postxid: str
     misc: str
-    leverage: float
-    margin: float
+    leverage: Decimal  # Changed if needed for precise calculations
+    margin: Decimal    # Changed if needed for precise calculations
     type: str
  
 @dataclass
@@ -31,12 +32,12 @@ class Pair:
     
 @dataclass
 class ExchangeRate:
-    rateDate: datetime.date # date of the exchange rate (might be the same date or earlier as per logic T minus 0,1,2,3,...n)
-    rate: float    
+    rateDate: datetime.date
+    rate: Decimal   # Should also be Decimal
     base_currency: str 
     quote_currency: str # for Poland is PLN (for CryptoTaxPL this should be always PLN)  
     
 @dataclass
 class TransactionForTax:
     transaction: Transaction
-    tax_exchange_rate: ExchangeRate  
+    tax_exchange_rate: ExchangeRate
